+++ source/Patches/CrewmateRoles/MedicMod/DeadBody.cs
@@ -1,6 +1,8 @@
 ﻿using System;
-using TownOfUs.Roles;
 using System.Collections.Generic;
+using System.Linq;
+using HarmonyLib;
+using UnityEngine;
 
 namespace TownOfUs.CrewmateRoles.MedicMod
 {
@@ -19,6 +21,45 @@ namespace TownOfUs.CrewmateRoles.MedicMod
         public PlayerControl Body { get; set; }
         public float KillAge { get; set; }
 
+        // move the colorType dictionnary out the function to use in other method and initialize once
+        private static Dictionary<int, string> ColorType = new Dictionary<int, string>
+        {
+            {0, "darker"},// red
+            {1, "darker"},// blue
+            {2, "darker"},// green
+            {3, "lighter"},// pink
+            {4, "lighter"},// orange
+            {5, "lighter"},// yellow
+            {6, "darker"},// black
+            {7, "lighter"},// white
+            {8, "darker"},// purple
+            {9, "darker"},// brown
+            {10, "lighter"},// cyan
+            {11, "lighter"},// lime
+            {12, "darker"},// maroon
+            {13, "lighter"},// rose
+            {14, "lighter"},// banana
+            {15, "darker"},// gray
+            {16, "darker"},// tan
+            {17, "lighter"},// coral
+            {18, "darker"},// watermelon
+            {19, "darker"},// chocolate
+            {20, "lighter"},// sky blue
+            {21, "darker"},// beige
+            {22, "lighter"},// hot pink
+            {23, "lighter"},// turquoise
+            {24, "lighter"},// lilac
+            {25, "darker"},// rainbow
+            {26, "lighter"},// azure
+            {27, "darker"},// Panda
+        };
+
+        //function to get the color type of a player
+        public static String GetColorType(PlayerControl player)
+        {
+            return ColorType[player.Data.ColorId];
+        }
+
         public static string ParseBodyReport(BodyReport br)
         {
             //System.Console.WriteLine(br.KillAge);
@@ -34,10 +75,36 @@ namespace TownOfUs.CrewmateRoles.MedicMod
                 return
                     $"Body Report: The killer appears to be {br.Killer.Data.PlayerName}! (Killed {Math.Round(br.KillAge / 1000)}s ago)";
 
-            //deplace the colortype list into Role.cs and call back here
-            var typeOfColor = Role.HudManager_Update.GetColorType(br.Killer);
+            //Call the GetColorType method
+            var typeOfColor = GetColorType(br.Killer);
             return
                 $"Body Report: The killer appears to be a {typeOfColor} color. (Killed {Math.Round(br.KillAge / 1000)}s ago)";
         }
     }
+
+    //Update meeting to clorize the darker color name in black
+    [HarmonyPriority(Priority.Last)]
+    [HarmonyPatch(typeof(HudManager), nameof(HudManager.Update))]
+    public static class HudManager_Update
+    {
+        private static void PostFix(MeetingHud __instance)
+        {
+            var localPlayer = PlayerControl.LocalPlayer;
+            if (MeetingHud.Instance == null) return;
+            if (localPlayer.Data.IsDead) return;
+            if (!localPlayer.Is(RoleEnum.Medic)) return;
+
+            foreach (var playerState in __instance.playerStates)
+            {
+                var player = PlayerControl.AllPlayerControls.ToArray()
+                    .FirstOrDefault(x => x.PlayerId == playerState.TargetPlayerId);
+
+                if (localPlayer != player && player != null && player.Data != null)
+                {
+                    if (BodyReport.GetColorType(player) == "darker")
+                        playerState.NameText.color = Color.black;
+                }
+            }
+        }
+    }                      
 }
